Rename refusjonsflyten
