+++ src/main/java/no/nav/familie/inntektsmelding/refusjonomsorgsdager/rest/ArbeidsforholdDto.java
@@ -1,4 +1,4 @@
-package no.nav.familie.inntektsmelding.refusjonomsorgsdagerarbeidsgiver.rest;
+package no.nav.familie.inntektsmelding.refusjonomsorgsdager.rest;
 
 public record ArbeidsforholdDto(String organisasjonsnummer, String arbeidsforholdId) {
 }
