+++ src/main/java/no/nav/familie/inntektsmelding/refusjonomsorgsdager/rest/InnloggetBrukerDto.java
@@ -1,4 +1,4 @@
-package no.nav.familie.inntektsmelding.refusjonomsorgsdagerarbeidsgiver.rest;
+package no.nav.familie.inntektsmelding.refusjonomsorgsdager.rest;
 
 import jakarta.validation.constraints.NotNull;
