+++ src/main/java/no/nav/familie/inntektsmelding/server/app/api/ApiConfig.java
@@ -17,7 +17,7 @@ import org.slf4j.LoggerFactory;
 import no.nav.familie.inntektsmelding.forespørsel.rest.ForespørselRest;
 import no.nav.familie.inntektsmelding.imdialog.rest.InntektsmeldingDialogRest;
 import no.nav.familie.inntektsmelding.overstyring.InntektsmeldingFpsakRest;
-import no.nav.familie.inntektsmelding.refusjonomsorgsdagerarbeidsgiver.rest.RefusjonOmsorgsdagerArbeidsgiverRest;
+import no.nav.familie.inntektsmelding.refusjonomsorgsdager.rest.RefusjonOmsorgsdagerRest;
 import no.nav.familie.inntektsmelding.server.auth.AutentiseringFilter;
 import no.nav.familie.inntektsmelding.server.exceptions.ConstraintViolationMapper;
 import no.nav.familie.inntektsmelding.server.exceptions.GeneralRestExceptionMapper;
@@ -57,7 +57,7 @@ public class ApiConfig extends ResourceConfig {
         return Set.of(ForespørselRest.class,
             InntektsmeldingDialogRest.class,
             InntektsmeldingFpsakRest.class,
-            RefusjonOmsorgsdagerArbeidsgiverRest.class,
+            RefusjonOmsorgsdagerRest.class,
             ArbeidsgiverinitiertDialogRest.class);
     }
